+++ flutter_inappwebview_platform_interface/lib/src/in_app_webview/platform_webview.g.dart
@@ -564,7 +564,6 @@ enum PlatformWebViewCreationParamsProperty {
   ///- [printJobController]:
   ///    - Android WebView
   ///    - iOS WKWebView
-  ///    - macOS WKWebView
   ///
   ///Use the [PlatformWebViewCreationParams.isPropertySupported] method to check if this property is supported at runtime.
   ///{@endtemplate}