+++ flutter_inappwebview_platform_interface/lib/src/types/permission_request.g.dart
@@ -43,11 +43,14 @@ class PermissionRequest {
     if (map['resources'] != null) {
       instance.resources = List<PermissionResourceType>.from(
         map['resources'].map(
-          (e) => switch (enumMethod ?? EnumMethod.nativeValue) {
-            EnumMethod.nativeValue => PermissionResourceType.fromNativeValue(e),
-            EnumMethod.value => PermissionResourceType.fromValue(e),
-            EnumMethod.name => PermissionResourceType.byName(e),
-          }!,
+          (e) =>
+              switch (enumMethod ?? EnumMethod.nativeValue) {
+                EnumMethod.nativeValue =>
+                  PermissionResourceType.fromNativeValue(e),
+                EnumMethod.value => PermissionResourceType.fromValue(e),
+                EnumMethod.name => PermissionResourceType.byName(e),
+              } ??
+              PermissionResourceType.UNKNOWN,
         ),
       );
     }